requirements.txt
motor==2.1.0
requests==2.22
aiofiles==0.5.0
redis==3.4.0
aiohttp==3.6.2
python_dateutil==2.8.1
